Split CommandManager into domain-specific classes (1.12)
Extract DailyCommands, TaskCommands, ReportCommands, SearchCommands,
AdminCommands from the 933-line God Object. CommandManager becomes a
thin facade with the same public interface so no CLI changes are needed.

Update test patches to target the correct domain modules instead of
the old monolithic commands module.
